move the clients module to the models package
