+++ api/routes/handlers/attachment_handlers.py
@@ -1,5 +1,5 @@
 from fastapi import Depends
-from clients import AWSClients, get_aws_clients
+from models.clients import AWSClients, get_aws_clients
 
 # =========================
 # |  ATTACHMENT HANDLERS  |
